Adds code style checker for Bash scripts (#751)

+++ scripts/code-style-kotlin.sh
@@ -17,18 +17,16 @@ readonly ktlint_releases="https://github.com/pinterest/ktlint/releases/download"
 
 run_ktlint() {
 
-	if ! test -f "$ktlint_bin"; then
-	    echo -e "${cyan}• Installing ktlint (v$ktlint_version)${normal}"
-	    mkdir -p "$ktlint_home"
-	    curl -sSL "$ktlint_releases/$ktlint_version/ktlint" -o "$ktlint_bin"
-	    chmod +x "$ktlint_bin"
-	fi
+    if ! test -f "$ktlint_bin"; then
+        echo -e "${cyan}• Installing ktlint (v$ktlint_version)${normal}"
+        mkdir -p "$ktlint_home"
+        curl -sSL "$ktlint_releases/$ktlint_version/ktlint" -o "$ktlint_bin"
+        chmod +x "$ktlint_bin"
+    fi
 
-	echo -e "${cyan}• Running ktlint (v$ktlint_version)${normal}"
-	
-	"$ktlint_bin" --reporter=plain?group_by_file --android
-
-	echo -e "${cyan}• No issues found with ktlint${normal}"
+    echo -e "${cyan}• Running ktlint (v$ktlint_version)${normal}"
+    "$ktlint_bin" --reporter=plain?group_by_file --android
+    echo -e "${cyan}• No issues found with ktlint${normal}"
 }
 
 readonly detekt_version="1.21.0"
@@ -38,23 +36,21 @@ readonly detekt_releases="https://github.com/detekt/detekt/releases/download"
 
 run_detekt() {
 
-	if ! test -f "$detekt_bin"; then
-	    echo -e "${cyan}• Installing detekt (v$detekt_version)${normal}"
-	    mkdir -p "$detekt_home"
-	    
-	    local detekt_zip="detekt-cli-$detekt_version.zip"
-	    
-	    rm -rf "$detekt_home/$detekt_zip"
-	    curl -sSL "$detekt_releases/v$detekt_version/$detekt_zip" -o "$cache_dir/$detekt_zip"
-	    unzip "$cache_dir/$detekt_zip" -d "$detekt_home" >/dev/null 2>&1
-	    chmod +x "$detekt_bin"
-	fi
+    if ! test -f "$detekt_bin"; then
+        echo -e "${cyan}• Installing detekt (v$detekt_version)${normal}"
+        mkdir -p "$detekt_home"
 
-	echo -e "${cyan}• Running detekt (v$detekt_version)${normal}"
-	
-	"$detekt_bin" -c .config/detekt.yml --build-upon-default-config
+        local detekt_zip="detekt-cli-$detekt_version.zip"
 
-	echo -e "${cyan}• No issues found with detekt${normal}"
+        rm -rf "${detekt_home:?}/${detekt_zip:?}"
+        curl -sSL "$detekt_releases/v$detekt_version/$detekt_zip" -o "$cache_dir/$detekt_zip"
+        unzip "$cache_dir/$detekt_zip" -d "$detekt_home" >/dev/null 2>&1
+        chmod +x "$detekt_bin"
+    fi
+
+    echo -e "${cyan}• Running detekt (v$detekt_version)${normal}"
+    "$detekt_bin" -c .config/detekt.yml --build-upon-default-config
+    echo -e "${cyan}• No issues found with detekt${normal}"
 }
 
 readonly help_ktlint="Run ktlint static analyser for this project"
@@ -64,11 +60,10 @@ readonly help_all="Run ktlint and detekt for this project"
 usage() {
     echo "Usage instructions:"
     echo
-    echo -e " ‣ static-analysis.sh ${cyan}ktlint${normal}  $help_ktlint"
-    echo -e " ‣ static-analysis.sh ${cyan}detekt${normal}  $help_detekt"    
-    echo -e " ‣ static-analysis.sh ${cyan}all${normal}     $help_all"   
+    echo -e "‣ static-analysis.sh ${cyan}ktlint${normal}  $help_ktlint"
+    echo -e "‣ static-analysis.sh ${cyan}detekt${normal}  $help_detekt"
+    echo -e "‣ static-analysis.sh ${cyan}all${normal}     $help_all"
 }
-
 
 readonly what="$1"
 
@@ -90,7 +85,7 @@ main() {
         run_ktlint
         echo
         run_detekt
-        ;;        
+        ;;
     *)
         echo -e "Error: Invalid linter → ${cyan}$what${normal}"
         usage
